fix mypy issues
